requirements.txt
Flask==2.3.2
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
pandas==2.0.3
